% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_ca_simple_effect}
\alias{anova_ca_simple_effect}
\title{anova_cca/rda_simple_effect}
\usage{
anova_ca_simple_effect(
  data,
  env,
  fun = c("RDA", "CCA"),
  permutations = permute::how(nperm = 999)
)
}
\arguments{
\item{data}{comm data}

\item{env}{env data}

\item{fun}{"RDA" or "CCA"}

\item{permutations}{permutation test form.}
}
\value{
The effect of each factor alone.
}
\description{
anova_cca/rda_simple_effect
}
\examples{
library(vegan)
data(dune)
data(dune.env)
anova_ca_simple_effect(dune, dune.env, "RDA", 499)
}
