% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_dist}
\alias{beta_dist}
\title{beta}
\usage{
beta_dist(otu, tree)
}
\arguments{
\item{otu}{otu table.}

\item{tree}{phylogene tree}
}
\value{
beta dist
}
\description{
compute some dists.
Plot the PCA PCoa NMDS.
}
\examples{
data(its)
tree <- ape::rtree(nrow(its), rooted = TRUE, tip.label = rownames(its))
dists <- beta_dist(its, tree)
pcoa_plot(dists[[3]])
}
