% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{get_corr}
\alias{get_corr}
\alias{get_corr_sig}
\title{corr p
Simplified digital display.}
\usage{
get_corr(df, cut = 0.3)

get_corr_sig(
  df,
  p_value,
  p_adj_method = "none",
  sep = "\\n",
  rm_no_sig = TRUE,
  rm_1 = TRUE,
  level = 3
)
}
\arguments{
\item{df}{your data}

\item{cut}{Numbers less than this value are omitted.}

\item{p_value}{p_value table}

\item{p_adj_method}{?stats::p.adjust.methods}

\item{sep}{Separator of r and *.}

\item{rm_no_sig}{Logical value, whether to remove non-significant values.}

\item{rm_1}{Logical value, whether to remove self-correlation values.}

\item{level}{significance level}
}
\value{
short cor R.

paste r and *.
}
\description{
corr p
Simplified digital display.
}
\examples{
get_corr(c(0.1, 0.5, 0.6), cut = 0.2)
get_corr_sig(c(0.1, 0.5, 0.6), c(0.1, 0.04, 0.001))
}
