% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ext}
\alias{get_ext}
\alias{get_sig}
\title{Easy get}
\usage{
get_ext(filename, lower = TRUE)

get_sig(df, level = 3, na = FALSE)
}
\arguments{
\item{filename}{file path}

\item{lower}{All lowercase}

\item{df}{your data}

\item{level}{significance level}

\item{na}{Whether to display negative numbers as NA.}
}
\value{
Convert significance numbers to ***.
}
\description{
Easy get
}
\examples{
get_ext("123.txt")
get_sig(c(-1,0.00001, 0.0001, 0.001, 0.01, 0.04), level = 5, na = FALSE)
get_sig(c(-1,0.00001, 0.0001, 0.001, 0.01, 0.04), level = 5, na = TRUE)
}
