% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_bw2}
\alias{theme_bw2}
\alias{theme_classic2}
\title{ggTheme}
\usage{
theme_bw2(font_size = 10, font = "sans", line_size = 0.5)

theme_classic2(font_size = 10, font = "sans", line_size = 0.5)
}
\arguments{
\item{font_size}{font_size}

\item{font}{font: sans = Arial, serif = Times New Roman, any other fonts}

\item{line_size}{line_size}
}
\description{
Custom ggTheme
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, disp, color = cyl)) +
    geom_point()
p + theme_bw2(font_size = 15, line_size = 2)
p + theme_classic2(font_size = 15, line_size = 2)
}
