% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{yyload}
\alias{yyload}
\alias{yyunload}
\title{yyload}
\usage{
yyload(..., names = NULL, show.conflict = TRUE)

yyunload(..., names = NULL, uninstall = FALSE)
}
\arguments{
\item{...}{names of pkgs, without the quotes.}

\item{names}{names of pkgs, with the quotes.}

\item{show.conflict}{Whether to display conflict information.}

\item{uninstall}{Logical value, whether to uninstall the installation package.}
}
\description{
library packages.
}
\examples{
yyload(ggrepel)
yyunload(ggrepel)
}
