% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_tools.R
\name{lst_mv}
\alias{lst_mv}
\title{Var tools}
\usage{
lst_mv(..., names = NULL, rm = TRUE)
}
\arguments{
\item{...}{vars.}

\item{names}{character, the names of vars. eg: ls(pattern = "aa").}

\item{rm}{logical, whether to delete the variable.}
}
\value{
new list
}
\description{
Move the var to a new specified list.
}
\examples{
a01 <- 1
a02 <- 2
b03 <- 1
all <- lst_mv(b03, names=ls(pattern = "a0"))

}
