% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{rm_NAs}
\alias{rm_NAs}
\title{Remove all na cols, or rows}
\usage{
rm_NAs(df, margin = 0)
}
\arguments{
\item{df}{your data}

\item{margin}{1 means rm all NA rows, 2 means rm all NA cols,
0 means all rows and cols.}
}
\value{
new data
}
\description{
Remove all na cols, or rows
}
\examples{
df <- 1:20 \%>\% matrix(nrow = 5)
df[2, 4] <- NA
df[5, 2] <- NA
df[1,] <- NA
df[, 3] <- NA
rownames(df) <- letters[1:5]
colnames(df) <- letters[6:9]
rm_NAs(df, margin=1)
rm_NAs(df, margin=2)
rm_NAs(df, margin=0)


}
