% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R, R/pca.R
\name{alpha_index}
\alias{alpha_index}
\alias{pca_plot}
\alias{pcoa_plot}
\alias{nmds_plot}
\title{alpha}
\usage{
alpha_index(otu, tree)

pca_plot(
  df,
  group,
  scale = FALSE,
  axis = c(1, 2),
  label = TRUE,
  legend = FALSE,
  point_size = 2,
  font = "sans",
  font_size = 10
)

pcoa_plot(
  dist,
  group,
  axis = c(1, 2),
  label = TRUE,
  legend = FALSE,
  point_size = 2,
  font = "sans",
  font_size = 10,
  ...
)

nmds_plot(
  dist,
  group,
  label = TRUE,
  legend = FALSE,
  point_size = 2,
  font = "sans",
  font_size = 10,
  ...
)
}
\arguments{
\item{otu}{otu table}

\item{tree}{tree}

\item{df}{df}

\item{group}{group}

\item{scale}{TRUE of FALSE. Scale species to unit variance
(like correlations).}

\item{axis}{select the axis to plot}

\item{label}{TRUE of FALSE. Whether to draw a label}

\item{legend}{TRUE of FALSE. Whether to draw a legend}

\item{point_size}{point size}

\item{font}{sans = Arial, serif = Times New Roman}

\item{font_size}{font size}

\item{dist}{dist.}

\item{...}{Other params from ?ape::pcoa and ?vegan::metaMDS}
}
\value{
alpha dist
}
\description{
compute some indexs.

Plot the PCA PCoa NMDS.
}
\examples{
data(its)
indexs <- alpha_index(its)
indexs
data(its)
otu_t <- t(its)
pca_plot(otu_t)
bray_dist <- vegan::vegdist(otu_t)
pcoa_plot(bray_dist)
nmds_plot(bray_dist)
}
