% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rarefy_table}
\alias{rarefy_table}
\alias{rarefy_plot}
\title{Rarefy}
\usage{
rarefy_table(x, step = 10, rep = 5, depth)

rarefy_plot(
  df,
  group,
  ylab = "Observed features",
  errorbar = FALSE,
  legend = TRUE,
  font = "sans",
  font_size = 10,
  id_name = "id",
  depth_name = "depth",
  value_name = "value"
)
}
\arguments{
\item{x}{the features(otu/asv) table,
The sample should be in column,
the features(otu/asv) should be in row.}

\item{step}{steps}

\item{rep}{replace}

\item{depth}{max sequencing depth}

\item{df}{one of result in rarefy_table,or other table}

\item{group}{group}

\item{ylab}{"Observed features", "Shannon"}

\item{errorbar}{TRUE, FALSE}

\item{legend}{TRUE, FALSE}

\item{font}{"sans" = Arial, "serif" = Times New Roman ,any fonts}

\item{font_size}{size_name of  the df}

\item{id_name}{id_name of the df}

\item{depth_name}{depth_name of  the df}

\item{value_name}{value_name of  the df}
}
\value{
rarefy lists with "Observed features" and "Shannon" indexs
}
\description{
Rarefy the otu/asv table, with "Observed features" and "Shannon" indexs.
Plot the rearrarefaction curve.
}
\examples{
data(its)
its_t <- t(its)
rt <- rarefy_table(its_t, step = 10, rep = 6)
rarefy_plot(rt[[1]])
rarefy_plot(rt[[2]], ylab = "Shannon")
}
