% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{yyread}
\alias{yyread}
\alias{yywrite}
\alias{yysave}
\title{Read Write Save}
\usage{
yyread(
  filename,
  support_excel = TRUE,
  excel_col_names = TRUE,
  excel_col_types = NULL,
  ...
)

yywrite(x, file = "", sep, eol = c("\\n", "\\r\\n", "\\r"))

yysave(
  plot = last_plot(),
  filename,
  path = "plot_dir",
  width = 8,
  height = 7.5,
  units = "cm",
  dpi = 1200,
  compression = "lzw",
  ...
)
}
\arguments{
\item{filename}{filename}

\item{support_excel}{TRUE or FALSE}

\item{excel_col_names}{TRUE to use the first row as column names,
FALSE to get default names, or a character vector giving a name for each column.
If user provides col_types as a vector, col_names can have one entry per column,
i.e. have the same length as col_types, or one entry per unskipped column.}

\item{excel_col_types}{Either NULL to guess all from the spreadsheet
or a character vector containing one entry per column from these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list".
If exactly one col_type is specified, it will be recycled.
The content of a cell in a skipped column is never read
and that column will not appear in the data frame output.
A list cell loads a column as a list of length 1 vectors,
which are typed using the type guessing logic from col_types = NULL,
but on a cell-by-cell basis.}

\item{...}{others values from ?readxl::read_excel, ?data.table::fread and ?ggsave.}

\item{x}{df, matrix}

\item{file}{Output file name. "" indicates output to the console.}

\item{sep}{The separator between columns.
Default "," for "csv" file; "\\t" otherswise.}

\item{eol}{Default "\\n", you can also use "\\r\\n" and "\\r".}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{path}{Path to the xls/xlsx file.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{compression}{the type of compression to be used.}
}
\description{
Easy read, easy write, easy save.
Function from data.table, readxl and ggplot.
}
\examples{
?data.table::fread
?readxl::read_excel
?data.table::fwrite
?ggplot2::ggsave
}
