% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_attach.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%T>\%}
\alias{read.newick}
\alias{clipboard}
\alias{\%do\%}
\alias{foreach}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{foreach}{\code{\link[foreach:foreach]{\%do\%}}, \code{\link[foreach]{foreach}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{readr}{\code{\link[readr]{clipboard}}}

  \item{treeio}{\code{\link[treeio]{read.newick}}}
}}

