% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_tools.R
\name{rename_var}
\alias{rename_var}
\title{Rename var}
\usage{
rename_var(old_var_name, new_var_name)
}
\arguments{
\item{old_var_name}{The name of the old variable.}

\item{new_var_name}{The name of the new variable.}
}
\value{
no return
}
\description{
Rename var
}
\examples{
a=1
rename_var("a", "c")
c

}
