% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{rep_run}
\alias{rep_run}
\title{repeat run}
\usage{
rep_run(exp, times = 3, .combine = list, ...)
}
\arguments{
\item{exp}{A task enclosed in quotes ready to run.}

\item{times}{times of repetition.}

\item{.combine}{function that is used to process the tasks results as
they generated.  This can be specified as either a function or
a non-empty character string naming the function.
Specifying 'c' is useful for concatenating the results into
a vector, for example.  The values 'cbind' and 'rbind' can combine
vectors into a matrix.  The values '+' and '*' can be used to
process numeric data.
By default, the results are returned in a list.}

\item{...}{Other arguments from foreach::foreach}
}
\value{
repeat results
}
\description{
repeat run
}
\examples{
rep_run("sum(1:10)")
"runif(10)" \%>\% rep_run(5, .combine = rbind)
}
