% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_num.R
\name{rnorm_std}
\alias{rnorm_std}
\title{generate radam numbers}
\usage{
rnorm_std(n = 5, mean = 0, sd = 1, digits = 4)
}
\arguments{
\item{n}{A positive number, the number of items to generated.}

\item{mean}{mean}

\item{sd}{sd}

\item{digits}{To the decimal places.}
}
\value{
numbers.
}
\description{
generate radam numbers
}
\examples{
rnorm_std(10)
}
