% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{get_sig}
\alias{get_sig}
\alias{sig_label}
\title{significance
Get simplified correlation coefficients and significance stars}
\usage{
get_sig(p_value, level = 3, p_adj_method = "none")

sig_label(
  formula,
  data,
  test = c("wilcox_test", "t_test", "dunn_test", "tukey_hsd", "HSD.test",
    "duncan.test", "LSD.test", "SNK.test"),
  p.adjust.method = "none",
  alpha = 0.05,
  quantile_function = "qt"
)
}
\arguments{
\item{p_value}{p_value table}

\item{level}{significance level}

\item{p_adj_method}{?stats::p.adjust.methods}

\item{formula}{a formula of the form x ~ group where x is a numeric variable
giving the data values and group is a factor with one or multiple levels
giving the corresponding groups. For example, formula = TP53 ~ cancer_group.}

\item{data}{a data.frame containing the variables in the formula.}

\item{test}{the name of the statistical test that is applied to the values.
"wilcox_test", "t_test", "dunn_test", "tukey_hsd" comes from rstatix,
"HSD.test", "duncan.test", "LSD.test", "SNK.test" comes from agricolae.}

\item{p.adjust.method}{method to adjust p values for multiple comparisons.
This parameter only works with the test method from rstatix.}

\item{alpha}{Significant level}

\item{quantile_function}{Quantile function to compute the confidence interval. Default is \code{qt} which means the \eqn{t} distribution and \code{qnorm} means \eqn{norm} distribution.}
}
\value{
Convert significance numbers to ***.

Difference analysis results and significance labeling.
}
\description{
significance
Get simplified correlation coefficients and significance stars

significance
Get difference analysis results and significance labeling.
The core source of this function comes from \code{rstatix} and \code{agricolae}.
}
\examples{
get_sig(c(0.6, 0.00001, 0.0001, 0.001, 0.01, 0.04), level = 5)
sig_label(disp ~ gear, mtcars)
}
