% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_tools.R
\name{yylst}
\alias{yylst}
\title{Var tools}
\usage{
yylst(ls)
}
\arguments{
\item{ls}{ls(); ls(pattern = "aa")}
}
\value{
new list
}
\description{
Move the var to a new specified list.
}
\examples{
a01 <- 1
a02 <- 2
a_all <- yylst(ls(pattern = "a0"))

}
