% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{yysave}
\alias{yysave}
\alias{yydev}
\alias{yyexport}
\title{yyplot}
\usage{
yysave(
  plot = last_plot(),
  filename = "Rplot\%03d.tif",
  width = 8,
  height = 6,
  path = "./plot_out",
  units = c("cm", "in"),
  dpi = 600,
  compression = "lzw",
  ...
)

yydev(
  filename = "Rplot\%03d.tif",
  width = 8,
  height = 6,
  path = "./plot_out",
  dpi = 600,
  verbose = TRUE,
  ...
)

yyexport(
  plot = last_plot(),
  filename = "Rplot\%03d.tif",
  width = 8,
  height = 6,
  path = "./plot_out",
  dpi = 600,
  ...
)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{filename}{File name to create on disk.}

\item{width}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{height}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{compression}{the type of compression to be used.}

\item{...}{others values from svglite::svglite, grDevices::cairo_pdf,
grDevices::tiff, grDevices::jpeg, grDevices::png.}

\item{verbose}{logical, whether to display prompts.}
}
\description{
Save ggplot object or create plot device.
}
\examples{
?ggplot2::ggsave
\dontrun{
## build device
yydev("tmp001.tif")
ggplot(mtcars, aes(mpg, wt)) +
  geom_point()
dev.off()

## Export the ggplot image directly.
ggplot(mtcars, aes(mpg, wt)) +
  geom_point()
yyexport("tmp002.pdf")
}

}
