% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{yyread}
\alias{yyread}
\title{Easily read files}
\usage{
yyread(
  filename,
  rownames = FALSE,
  header = TRUE,
  method = c("vroom", "fread", "excel"),
  sheet = NULL,
  col_types = NULL,
  encoding = "unknown",
  ...
)
}
\arguments{
\item{filename}{filename}

\item{rownames}{Logical; whether to use the first column as the row names.}

\item{header}{Logical; whether to use the first row as the col names.}

\item{method}{c("vroom", "fread", "excel"). The function used to read the file, the default is "vroom".
It is strongly recommended to use TSV or CSV files rather than Excel files}

\item{sheet}{Excel option; either a string (the name of a sheet),
or an integer (the position of the sheet).}

\item{col_types}{Excel option; NULL to guess;
A character vector containing these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list".
If exactly one col_type is specified, it will be recycled.
The content of a cell in a skipped column is never read}

\item{encoding}{Default is "unknown". Other possible options are "UTF-8" and "Latin-1".
and that column will not appear in the data frame output.}

\item{...}{Others params from \code{\link[vroom:vroom]{vroom::vroom()}}, \code{\link[data.table:fread]{data.table::fread()}} or \code{\link[readxl:read_excel]{readxl::read_excel()}}.
Just like "num_threads" in vroom; "nThread" in data.table.}
}
\description{
Function from data.table, readxl.
}
\examples{
\dontrun{
yyread("a1.tsv")
yyread("a1.xls", method = "excel")
}
}
