% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{yywrite}
\alias{yywrite}
\title{Easily write files}
\usage{
yywrite(
  x,
  file = "",
  row.names = FALSE,
  col.names = TRUE,
  sep,
  eol = c("\\n", "\\r\\n", "\\r"),
  ...
)
}
\arguments{
\item{x}{Any \code{list} of same length vectors; e.g. \code{data.frame} and \code{data.table}. If \code{matrix}, it gets internally coerced to \code{data.table} preserving col names but not row names}

\item{file}{Output file name. \code{""} indicates output to the console. }

\item{row.names}{Should row names be written? For compatibility with \code{data.frame} and \code{write.csv} since \code{data.table} never has row names. Hence default \code{FALSE} unlike \code{write.csv}.}

\item{col.names}{Should the column names (header row) be written? The default is \code{TRUE} for new files and when overwriting existing files (\code{append=FALSE}). Otherwise, the default is \code{FALSE} to prevent column names appearing again mid-file when stacking a set of \code{data.table}s or appending rows to the end of a file.}

\item{sep}{The separator between columns.
Default "," for "csv" file; "\\t" otherswise.}

\item{eol}{Default "\\n", you can also use "\\r\\n" and "\\r".}

\item{...}{others values from ?data.table::fwrite.}
}
\description{
Function from data.table.
}
\examples{
\dontrun{
a1 <- data.frame(matrix(1:9, 3))
yywrite(a1, "a1.tsv")
yywrite(a1, "a1.csv")
rownames(a1) <- letters[1:3]
yywrite(a1, "a1_name.tsv", row.names = TRUE)
yywrite(a1, "a1_name_no_colname.tsv", row.names = TRUE, col.names = FALSE)
}
}
