% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{get_sig}
\alias{get_sig}
\alias{get_corr}
\alias{get_corr_sig}
\title{corr p
Get simplified correlation coefficients and significance stars}
\usage{
get_sig(p_value, level = 3)

get_corr(df, cut = 0.3)

get_corr_sig(df, p_value, level = 3)
}
\arguments{
\item{p_value}{p_value table}

\item{level}{significance level}

\item{df}{your data}

\item{cut}{Numbers less than this value are omitted.}
}
\value{
Convert significance numbers to ***.
Simplified digital display.

paste r and *.
}
\description{
corr p
Get simplified correlation coefficients and significance stars
}
\examples{
get_sig(c(0.6, 0.00001, 0.0001, 0.001, 0.01, 0.04), level = 5)
get_corr(c(0.1, 0.5, 0.6), cut = 0.2)
get_corr_sig(c(0.1, 0.5, 0.6), c(0.1, 0.04, 0.001))
}
