% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ext}
\alias{get_ext}
\alias{get_sig}
\alias{get_corr}
\title{Easy get}
\usage{
get_ext(filename, lower = TRUE)

get_sig(df, level = 3, num = FALSE)

get_corr(df, cut = 0.3)
}
\arguments{
\item{filename}{file path}

\item{lower}{All lowercase}

\item{df}{your data}

\item{level}{significance level}

\item{num}{display numbers}

\item{cut}{Numbers less than this value are omitted.}
}
\value{
Convert significance numbers to ***.
Simplified digital display.
}
\description{
Easy get
}
\examples{
get_ext("123.txt")
get_sig(c(-1,0.00001, 0.0001, 0.001, 0.01, 0.04), level = 5)
get_corr(c(0.1, 0.5, 06))
}
