% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ext}
\alias{get_ext}
\title{Easy get}
\usage{
get_ext(filename, lower = TRUE)
}
\arguments{
\item{filename}{file path}

\item{lower}{All lowercase}
}
\description{
Easy get
}
\examples{
get_ext("123.txt")
}
