% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{yyinstall}
\alias{yyinstall}
\alias{yyuninstall}
\alias{yyload}
\alias{yyunload}
\title{yyload}
\usage{
yyinstall(
  ...,
  names = NULL,
  dependencies = NA,
  update = FALSE,
  version = BiocManager::version()
)

yyuninstall(..., names = NULL)

yyload(..., names = NULL, show.conflict = TRUE)

yyunload(..., names = NULL)
}
\arguments{
\item{...}{names of pkgs, without the quotes.}

\item{names}{names of pkgs, with the quotes.}

\item{dependencies}{logical indicating whether to also install
    uninstalled packages which these packages depend on/link
    to/import/suggest (and so on recursively).
    Not used if \code{repos = NULL}.
    Can also be a character vector, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.

    Only supported if \code{lib} is of length one (or missing),
    so it is unambiguous where to install the dependent packages.  If
    this is not the case it is ignored, with a warning.

    The default, \code{NA}, means
    \code{c("Depends", "Imports", "LinkingTo")}.

    \code{TRUE} means to use
    \code{c("Depends", "Imports", "LinkingTo", "Suggests")} for
    \code{pkgs} and
    \code{c("Depends", "Imports", "LinkingTo")} for added dependencies:
    this installs all the packages needed to run \code{pkgs}, their
    examples, tests and vignettes (if the package author specified them
    correctly).

    In all of these, \code{"LinkingTo"} is omitted for binary packages.
  }

\item{update}{\code{logical(1)}. When \code{FALSE}, \code{BiocManager::install()}
does not attempt to update old packages. When \code{TRUE}, update
old packages according to \code{ask}.}

\item{version}{\code{character(1)} \emph{Bioconductor} version to install,
e.g., \code{version = "3.8"}. The special symbol \code{version = "devel"}
installs the current 'development' version.}

\item{show.conflict}{Whether to display conflict information.}
}
\description{
library packages.
}
\examples{
\dontrun{
yyinstall(ggtext, "ggsci")
yyuninstall(ggtext, "ggsci")

## restart R
yyload(ggtext, "ggsci")
yyunload(ggtext, "ggsci")
}
}
