% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{maps_trans2gcj02}
\alias{maps_trans2gcj02}
\alias{maps_get_coords}
\alias{maps_get_address}
\alias{maps_get_elevation}
\title{Map tools}
\usage{
maps_trans2gcj02(coord, type = "baidu", key = amap_key, output = "JSON")

maps_get_coords(address, key = amap_key, output = "JSON")

maps_get_address(coord, key = amap_key, output = "JSON")

maps_get_elevation(coord, key = google_key)
}
\arguments{
\item{coord}{string, Latitude follows longitude, separated by commas.\\n
eg:"108.952560,34.204798"}

\item{type}{"baidu", "gps"}

\item{key}{amap key}

\item{output}{output type, default is JSON.}

\item{address}{The Chinese address}
}
\value{
coordinates of amap(GCJ-02). Or the Chinese address.
}
\description{
Coordinate transformation. Obtain the coordinate. Get the address.
}
\details{
Visit https://lbs.amap.com/api/webservice/summary
for more infomation.
NOTE: These functions only supoort chinese address.
}
\examples{
maps_get_coords("陕西师范大学")
maps_get_address("108.952560,34.204798")
maps_get_elevation("108.952560,34.204798")
maps_trans2gcj02("108.952560,34.204798", type = "baidu")
}
