% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R, R/pca.R
\name{alpha_index}
\alias{alpha_index}
\alias{pca_plot}
\alias{pcoa_plot}
\alias{nmds_plot}
\title{alpha}
\usage{
alpha_index(otu, tree)

pca_plot(
  df,
  group,
  scale = FALSE,
  axis = c(1, 2),
  label = TRUE,
  legend = FALSE,
  point_size = 2,
  font = "sans",
  font_size = 10,
  returns = c("plot", "data")
)

pcoa_plot(
  dist,
  group,
  axis = c(1, 2),
  label = TRUE,
  legend = FALSE,
  point_size = 2,
  font = "sans",
  font_size = 10,
  returns = c("plot", "data")
)

nmds_plot(
  dist_or_df,
  group,
  veddist_method = "bray",
  autotransform = FALSE,
  trace = FALSE,
  label = TRUE,
  legend = FALSE,
  point_size = 2,
  font = "sans",
  font_size = 10,
  returns = c("plot", "data")
)
}
\arguments{
\item{otu}{otu table}

\item{tree}{tree}

\item{df}{df}

\item{group}{group}

\item{scale}{TRUE of FALSE. Scale species to unit variance
(like correlations).}

\item{axis}{select the axis to plot}

\item{label}{TRUE of FALSE. Whether to draw a label}

\item{legend}{TRUE of FALSE. Whether to draw a legend}

\item{point_size}{point size}

\item{font}{sans = Arial, serif = Times New Roman}

\item{font_size}{font size}

\item{returns}{return plot or axis}

\item{dist}{dist}

\item{dist_or_df}{Community, otu , species data, Alternatively,
dissimilarities either as a dist structure .}

\item{veddist_method}{Dissimilarity index, partial match to
"manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski",
"jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao", "mahalanobis", "chisq" or "chord".}

\item{autotransform}{Use simple heuristics for possible data transformation
of typical community data (see below). If you do not have community data,
you should probably set autotransform = FALSE.}

\item{trace}{FALSE or unmber,
Trace the function; trace = 2 or higher will be more voluminous.}
}
\value{
alpha dist
}
\description{
compute some indexs.

Plot the PCA PCoa NMDS.
}
\examples{
data(its)
indexs <- alpha_index(its)
indexs

data(its)
otu_t <- t(its)
pca_plot(otu_t)
tree <- ape::rtree(nrow(its), rooted = TRUE, tip.label = rownames(its))
bray_dist <- vegan::vegdist(otu_t)
pcoa_plot(bray_dist)
nmds_plot(bray_dist)
nmds_plot(otu_t)

}
