% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{rarefy_otu}
\alias{rarefy_otu}
\title{rarefy}
\usage{
rarefy_otu(phy, step = 10, rep = 3, measures = c("Observed", "Shannon"))
}
\arguments{
\item{phy}{phyloseq}

\item{step}{steps}

\item{rep}{times of repetition}

\item{measures}{(Optional). Default is \code{NULL}, meaning that
all available alpha-diversity measures will be included.
Alternatively, you can specify one or more measures
as a character vector of measure names.
Values must be among those supported:
\code{c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher")}.}
}
\value{
rarefy table
}
\description{
Prepare data for rerafy curves.
}
\examples{
data(its)
phy <- phyloseq(otu_table(its, taxa_are_rows = TRUE))
rare <- rarefy_otu(phy)
x_min <- group_by(rare, id) \%>\%
    summarise(across(Depth, max)) \%>\%
    .[, "Depth"] \%>\%
    min()
ggplot(rare, aes(Depth, Observed, group = id, color = id)) +
    geom_vline(xintercept = x_min, linetype = "dashed", color = "grey60") +
    stat_summary(fun = "mean", geom = "smooth", size = .7) +
    stat_summary(
        fun.data = "mean_cl_normal", geom = "errorbar",
        size = .5, aes(width = max(Depth) / 30)) +
    theme_bw2(15) +
    theme(legend.justification = c(1, 0), legend.position = c(1, 0))
}
