% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_tools.R
\name{rename_var}
\alias{rename_var}
\title{Rename var}
\usage{
rename_var(...)
}
\arguments{
\item{...}{The old variable and new variable. Only two items are allowed.}
}
\value{
no return
}
\description{
Rename var
}
\examples{
a <- 1
rename_var(a, c)
c
}
