% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std_max}
\alias{std_max}
\alias{std_min}
\title{Max min standardization}
\usage{
std_max(x)

std_min(x)
}
\arguments{
\item{x}{Numbers.}
}
\value{
Standardized data
}
\description{
Max min standardization
}
\examples{
std_max(1:10)
std_min(1:10)
}
