% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{t.tbl_df}
\alias{t.tbl_df}
\alias{t.data.table}
\alias{t}
\title{Easily transpose tibble and data.frame and data.table}
\usage{
\method{t}{tbl_df}(x, colnames_to_columm = "col_id", column_to_colnames = 1, ...)

\method{t}{data.table}(x, colnames_to_columm = "col_id", column_to_colnames = 1, ...)

t(x, colnames_to_columm, column_to_colnames, ...)
}
\arguments{
\item{x}{an tibble or data.table}

\item{colnames_to_columm}{The name of the first column in the result containing the colnames of the input.}

\item{column_to_colnames}{The name or number of a column in the input to use as colnames of the output.}

\item{...}{Others params from \code{\link[data.table:transpose]{data.table::transpose()}}.}
}
\value{
a transpose tibble

a transpose data.table
}
\description{
Function from data.table.
}
\examples{
tibble::as_tibble(mtcars, rownames = "row_id") \%>\% t()
tibble::as_tibble(mtcars, rownames = "row_id") \%>\%
    t() \%>\%
    t()
tibble::as_tibble(mtcars, rownames = "row_id") \%>\%
    data.table::as.data.table() \%>\%
    t()

}
