% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{yycbind}
\alias{yycbind}
\title{yybind}
\usage{
yycbind(x, y, by)
}
\arguments{
\item{x}{the first df}

\item{y}{the second df}

\item{by}{item should be the rownames}
}
\value{
cbind df
}
\description{
Easy bind
}
\examples{
?cbind
?dplyr::bind_cols
}
