% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{yydev}
\alias{yydev}
\alias{yyexport}
\title{yyplot}
\usage{
yydev(
  filename = "Rplot\%03d.tif",
  width = 8,
  height = 6,
  path = "./plot_out",
  dpi = 600,
  verbose = TRUE,
  ...
)

yyexport(
  plot = last_plot(),
  filename = "Rplot\%03d.tif",
  width = 8,
  height = 6,
  path = "./plot_out",
  dpi = 600,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{width, height}{Plot size in cm. Defaults to the 8 x 6 cm.}

\item{path}{Path of the directory to save plot to.}

\item{dpi}{Plot resolution.}

\item{verbose}{logical, whether to display prompts.}

\item{...}{others values from svglite::svglite, grDevices::cairo_pdf,
grDevices::tiff, grDevices::jpeg, grDevices::png.}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save ggplot object or create plot device.
}
\examples{
\dontrun{
## build device
yydev("tmp001.tif")
ggplot(mtcars, aes(mpg, wt)) +
    geom_point()
dev.off()

## Export the ggplot image directly.
ggplot(mtcars, aes(mpg, wt)) +
    geom_point()
yyexport(filename = "tmp003.pdf")
}
}
